+++ CHANGELOG.md
@@ -16,6 +16,15 @@ limitations under the License.
 
 # Changelog
 
+## Unreleased
+
+- Fix: ModelManager does not raise ``tritonclient.grpc.InferenceServerException`` for ``stop`` method when HTTP endpoint is disabled in Triton configuration.
+
+[//]: <> (put here on external component update with short summary what change or link to changelog)
+
+- Version of [Triton Inference Server](https://github.com/triton-inference-server/) embedded in wheel: [2.42.0](https://github.com/triton-inference-server/server/releases/tag/v2.42.0)
+
+
 ## 0.5.1 (2024-02-09)
 
 - Fix: ModelClient does not raise `gevent.exceptions.InvalidThreadUseError` when destroyed in a different thread.